Adding package for Raduko * 2013.08 in lang/raduko-star.
Rakudo Perl 6, or simply Rakudo, is a compiler for the Perl 6 programming
language. It runs on the Parrot virtual machine.

The Rakudo Star distribution (fondly called Rakudo *) is the Rakudo
compiler plus a selection of useful Perl 6 modules, the most recent
incarnation of the "Using Perl 6" book, and other software that can
be used with the Rakudo compiler to enhance its utility.  Rakudo Star
is meant for early adopters who wish to explore what's possible with
Rakudo Perl 6 and provide feedback on what works, what doesn't, and
what else they would like to see included in the distribution.