Build form and menu support now.
Perl Curses can use system curses fine, but we need to hand edit the
c-config.h file to so as not to conflict the system forms.h file
with the perl one. This is as documented in the package.
NetBSD system curses does not support panels.

If we're using ncurses we also build support for panels.